--- a/herders/filters.py
+++ b/herders/filters.py
@@ -17,6 +17,7 @@ class MonsterInstanceFilter(django_filters.FilterSet):
     monster__skills__scaling_stats__pk = django_filters.ModelMultipleChoiceFilter(queryset=ScalingStat.objects.all(), to_field_name='pk', conjoined=True)
     monster__skills__skill_effect__pk = django_filters.ModelMultipleChoiceFilter(queryset=SkillEffect.objects.all(), method='filter_monster__skills__skill_effect__pk')
     monster__skills__cooltime = django_filters.CharFilter(method='filter_bypass')
+    monster__skills__hits = django_filters.CharFilter(method='filter_bypass')
     monster__skills__passive = django_filters.BooleanFilter(method='filter_bypass')
     monster__skills__aoe = django_filters.BooleanFilter(method='filter_bypass')
     effects_logic = django_filters.BooleanFilter(method='filter_bypass')
@@ -38,7 +39,6 @@ class MonsterInstanceFilter(django_filters.FilterSet):
             'monster__leader_skill__area': ['exact'],
             'monster__skills__skill_effect__pk': ['exact'],
             'monster__skills__scaling_stats__pk': ['exact'],
-            'monster__skills__hits': ['gte', 'lte'],
             'monster__skills__passive': ['exact'],
             'monster__skills__aoe': ['exact'],
             'effects_logic': ['exact'],
@@ -67,19 +67,24 @@ class MonsterInstanceFilter(django_filters.FilterSet):
     def filter_monster__skills__skill_effect__pk(self, queryset, name, value):
         old_filtering = self.form.cleaned_data.get('effects_logic', False)
         stat_scaling = self.form.cleaned_data.get('monster__skills__scaling_stats__pk', [])
-        cooltimes = self.form.cleaned_data.get('monster__skills__cooltime', [])
-        max_num_hits = self.form.cleaned_data.get('monster__skills__hits__lte', 99)
-        min_num_hits = self.form.cleaned_data.get('monster__skills__hits__gte', 0)
         passive = self.form.cleaned_data.get('monster__skills__passive', None)
         aoe = self.form.cleaned_data.get('monster__skills__aoe', None)
 
         try:
-            cooltimes = cooltimes.split(',')
-            min_cooltime = int(cooltimes[0])
-            max_cooltime = int(cooltimes[1])
+            [min_cooltime, max_cooltime] = self.form.cleaned_data['monster__skills__cooltime'].split(',')
+            min_cooltime = int(min_cooltime)
+            max_cooltime = int(max_cooltime)
         except:
-            min_cooltime = 0
-            max_cooltime = 999
+            min_cooltime = None
+            max_cooltime = None
+
+        try:
+            [min_num_hits, max_num_hits] = self.form.cleaned_data['monster__skills__hits'].split(',')
+            min_num_hits = int(min_num_hits)
+            max_num_hits = int(max_num_hits)
+        except:
+            min_num_hits = None
+            max_num_hits = None
 
         if old_filtering:
             # Filter if any skill on the monster has the designated fields
@@ -89,16 +94,25 @@ class MonsterInstanceFilter(django_filters.FilterSet):
             for pk in stat_scaling:
                 queryset = queryset.filter(monster__skills__scaling_stats=pk)
 
-            cooltime_filter = Q(monster__skills__cooltime__gte=min_cooltime) & Q(monster__skills__cooltime__lte=max_cooltime)
+            cooltime_filter = Q()
+            
+            if max_cooltime is not None and max_cooltime > 0:
+                cooltime_filter &= Q(monster__skills__cooltime__lte=max_cooltime)
+            
+            if min_cooltime is not None and min_cooltime > 0:
+                cooltime_filter &= Q(monster__skills__cooltime__gte=min_cooltime)
+
             if min_cooltime == 0 or max_cooltime == 0:
-                cooltime_filter = Q(monster__skills__cooltime__isnull=True) | cooltime_filter
+                cooltime_filter |= Q(monster__skills__cooltime__isnull=True)
 
-            queryset = queryset.filter(
-                cooltime_filter,
-                monster__skills__hits__lte=max_num_hits,
-                monster__skills__hits__gte=min_num_hits,
-            )
+            if cooltime_filter:
+                queryset = queryset.filter(cooltime_filter)
 
+            if max_num_hits:
+                queryset = queryset.filter(monster__skills__hits__lte=max_num_hits)
+
+            if min_num_hits:
+                queryset = queryset.filter(monster__skills__hits__gte=min_num_hits)
             if passive is not None:
                 queryset = queryset.filter(monster__skills__passive=passive)
 
@@ -111,6 +125,7 @@ class MonsterInstanceFilter(django_filters.FilterSet):
             # Filter effects based on effects of each individual skill. This ensures a monster will not show up unless it has
             # the desired effects on the same skill rather than across any skills.
             skills = Skill.objects.all()
+            skills_count = skills.count()
 
             for effect in value:
                 skills = skills.filter(skill_effect=effect)
@@ -118,21 +133,40 @@ class MonsterInstanceFilter(django_filters.FilterSet):
             for pk in stat_scaling:
                 skills = skills.filter(scaling_stats=pk)
 
-            cooltime_filter = Q(cooltime__gte=min_cooltime) & Q(cooltime__lte=max_cooltime)
-            if min_cooltime == 0 or max_cooltime == 0:
-                cooltime_filter = Q(cooltime__isnull=True) | cooltime_filter
+            cooltime_filter = Q()
 
-            skills = skills.filter(
-                cooltime_filter,
-                hits__lte=max_num_hits,
-                hits__gte=min_num_hits,
-            )
+            if max_cooltime is not None and max_cooltime > 0:
+                cooltime_filter &= Q(cooltime__lte=max_cooltime)
+            
+            if min_cooltime is not None and min_cooltime > 0:
+                cooltime_filter &= Q(cooltime__gte=min_cooltime)
+
+            if min_cooltime == 0 or max_cooltime == 0:
+                cooltime_filter |= Q(cooltime__isnull=True)
+
+            if cooltime_filter:
+                skills = skills.filter(cooltime_filter)
+            
+            hits_filter = Q()
+
+            if max_num_hits:
+                hits_filter &= Q(hits__lte=max_num_hits)
+
+            if min_num_hits:
+                hits_filter &= Q(hits__gte=min_num_hits)
+
+            if hits_filter:
+                skills = skills.filter(hits_filter)
 
             if passive is not None:
                 skills = skills.filter(passive=passive)
 
             if aoe is not None:
                 skills = skills.filter(aoe=aoe)
+
+            # no skill filters
+            if skills_count == skills.count():
+                return queryset.distinct()
 
             return queryset.filter(monster__skills__in=skills).distinct()