… api/templates/api/monster_instance/popover.html
… api/templates/api/monster_skills/popover.html
… api/templates/api/rune_instance/popover.html
… api/templates/api/rune_instance/thumbnail.html
… bestiary/templates/bestiary/detail_base.html
… bestiary/templates/bestiary/essence_fragment.html
… bestiary/templates/bestiary/inventory.html
… bestiary/templates/bestiary/monster/portrait_large.html
… bestiary/templates/bestiary/monster/portrait_small.html
… bestiary/templates/bestiary/monster/stars.html
… bestiary/templates/bestiary/monster_info.html
… bestiary/templates/bestiary/skills.html
… bestiary/templates/dungeons/base.html
… bestiary/templates/dungeons/detail/report_by_grade_summary_table.html
… data_log/templates/data_log/drops/item.html
… data_log/templates/data_log/reports/summon.html
… data_log/templates/data_log/summary_items.html
… data_log/templates/data_log/summary_monsters.html
… data_log/templates/data_log/summary_runes.html
… data_log/templates/data_log/summary_secret_dungeons.html
… herders/management/commands/propagate_test_data.py
… herders/static/herders/js/runes.js
… herders/templates/herders/monster_icon_button.html
… herders/templates/herders/profile/artifacts/artifact_panel_fragment.html
… herders/templates/herders/profile/base.html
… herders/templates/herders/profile/buildings/building_row_snippet.html
… herders/templates/herders/profile/buildings/inventory.html
… herders/templates/herders/profile/data_logs/drops/artifact_popover.html
… herders/templates/herders/profile/data_logs/drops/item_drops.html
… herders/templates/herders/profile/data_logs/drops/monster_drops.html