* Added Lazy loading for all images
* Cleaned up filter url (i.e. `name=&natural_stars=1,6&fusion_food=&skills__passive=&skills__aoe=&skills__cooltime=0,13&skills__hits=0,7&page=1&sort=` changed to `page=1`)
* Optimized Bestiary/Monsters Database calls

It also partially resolves given Feedback
> Whenever you click on bestiary, swarfarm load every monster before you can modify filter, making swarfam super slow especicaly on my phone, maybe, for the first loading, don't load monster and wait for the user to press apply to filter.

| Loading time [seconds]  | SWARFARM (before) | LOCAL (after) |
|---|---|-----|
| Bestiary  | 7  | 1  |  
| Monsters  | 6.5  | 3.5   |  
| Runes  | 10  | 8  |
| Artifacts | 1.5 | <0.5 |